gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6

# 추가 의존성 (필요시)
# redis==5.0.1
# aioredis==2.0.1